Add files via upload
This is a java project in which we guess a number and number is matched with computer generated number .If both the number (computer generated and guessed number ) are matched then the score will be display according to your attempts to guess the  number. I also add play again option in it and limiting the number of attempts.